Remove update resources tests, add to ruleshelpers test.    Add missing tests
