+++ packages/api/lib/orca.js
@@ -81,7 +81,7 @@ const getOrcaRecoveryStatusByGranuleId = async (granuleId) => {
 /**
  * add recovery status for each granule in the granule list response
  *
- * @param {Object} inputResponse - an elasticsearch reponse returned from granules query
+ * @param {Object} inputResponse - an elasticsearch response returned from granules query
  * @returns {Object} a copy of input response object where each granule
  *      has been updated to include orca recovery status
  */
